+++ sora-android-sdk/src/main/kotlin/jp/shiguredo/sora/sdk/channel/SoraMediaChannel.kt
@@ -69,6 +69,7 @@ import kotlin.concurrent.schedule
  * @param dataChannels connect メッセージに含める `data_channels`
  * @param bundleId connect メッセージに含める `bundle_id`
  * @param forwardingFilterOption 転送フィルター機能の設定
+ * @param forwardingFiltersOption リスト形式の転送フィルター機能の設定
  */
 class SoraMediaChannel @JvmOverloads constructor(
     private val context: Context,
@@ -86,7 +87,13 @@ class SoraMediaChannel @JvmOverloads constructor(
     ignoreDisconnectWebSocket: Boolean? = null,
     dataChannels: List<Map<String, Any>>? = null,
     private var bundleId: String? = null,
+    @Deprecated(
+        "この項目は 2025 年 12 月リリース予定の Sora にて廃止されます",
+        ReplaceWith("forwardingFiltersOption"),
+        DeprecationLevel.WARNING
+    )
     private val forwardingFilterOption: SoraForwardingFilterOption? = null,
+    private val forwardingFiltersOption: List<SoraForwardingFilterOption>? = null,
 ) {
     companion object {
         private val TAG = SoraMediaChannel::class.simpleName
@@ -672,6 +679,7 @@ class SoraMediaChannel @JvmOverloads constructor(
             |bundleId                   = ${this.bundleId}
             |signalingNotifyMetadata    = ${this.signalingNotifyMetadata}
             |forwardingFilter           = ${this.forwardingFilterOption}
+            |forwardingFilters           = ${this.forwardingFiltersOption}
             """.trimMargin()
         )
 
@@ -781,7 +789,8 @@ class SoraMediaChannel @JvmOverloads constructor(
             signalingNotifyMetadata = signalingNotifyMetadata,
             connectDataChannels = connectDataChannels,
             redirect = redirectLocation != null,
-            forwardingFilterOption = forwardingFilterOption
+            forwardingFilterOption = forwardingFilterOption,
+            forwardingFiltersOption = forwardingFiltersOption
         )
         signaling!!.connect()
     }